# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/habits/addHabit": {
            "post": {
                "description": "Add a habit with name and description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "habits"
                ],
                "summary": "Add a habit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Habit name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Habit description",
                        "name": "description",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesfully added habit",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Both 'name' and 'description' are required",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/habits/getHabits": {
            "get": {
                "description": "Get all habits from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "habits"
                ],
                "summary": "Get all habits",
                "responses": {
                    "200": {
                        "description": "List of habits",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Habit"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Habit": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}